Dev Env Setup for multiple version patch generation.
Bin Patches are now saved, as they are required for the mod build with multiple version support. They are saved under `src\main\generatedResrouces` directory.

ISSUE: Either Build is required to be runned twice, or generateBinPatches has to be launched before triggering the build. I wasn't able to get this to find freshly-generated files for whatever reason.

ISSUE: Dev Builds work only on a version of IC2 that is used to compile it. Dev Bin Patches are inserted to the jar root, and other patches are excluded.