Create task that patches compiled jar of Patched IC2 so it boots.
This task is meant to be launched only on prepareRuns, and it replaces entries complying to the regex rules with entries from the original IC2 Jar. Mainly used for assets, but also for networking classes.

Note: Literally just before this commit I realized there is old patch that just comments out entirety of IC2 Networking code and that was the cause of the crashes lmao, but this is still going to be useful for... some time until I decide to replace this with actual non-corrupted asset extraction and fix the networking base code so it's not entirely commented out and is stable enough to launch the game.